feat: 매장 위치 등록 실험실 구현
